Refactored some local state variables into global state
