+++ src/Store.js
@@ -1,11 +1,17 @@
 import React, { createContext, useReducer } from 'react';
 import Reducer from './Reducer';
 
+import Clips from './assets/Clips';
+import { shuffleArray } from './utils/MakeGame';
+
 const initialState = {
     asleep: false,
     awake: false,
     started: false,
     password: "",
+    index: 0,
+    answers: shuffleArray(Object.keys(Clips)),
+    questions: [],
     audio: "",
     playing: false,
 };
