+++ src/styles/Container.module.css
@@ -4,7 +4,6 @@
     flex-direction: column;
 	width: 100%;
     height: 100%;
-    /* row-gap: 3em; */
 	overflow: visible;
     align-items: center;
     justify-content: center;
